code refactor.

+++ src/main/java/ac/boar/anticheat/check/impl/velocity/VelocityA.java
@@ -4,7 +4,7 @@ import ac.boar.anticheat.check.api.CheckInfo;
 import ac.boar.anticheat.check.api.impl.OffsetHandlerCheck;
 import ac.boar.anticheat.prediction.engine.data.VectorType;
 import ac.boar.anticheat.user.api.BoarPlayer;
-import ac.boar.utils.math.Vec3d;
+import ac.boar.utils.math.Vec3f;
 import org.bukkit.Bukkit;
 
 import java.util.Iterator;
@@ -22,9 +22,9 @@ public class VelocityA extends OffsetHandlerCheck {
             return;
         }
 
-        Iterator<Map.Entry<Long, Vec3d>> iterator = player.queuedVelocities.entrySet().iterator();
+        Iterator<Map.Entry<Long, Vec3f>> iterator = player.queuedVelocities.entrySet().iterator();
 
-        Map.Entry<Long, Vec3d> entry;
+        Map.Entry<Long, Vec3f> entry;
         while (iterator.hasNext() && (entry = iterator.next()) != null) {
             if (player.lastReceivedId - entry.getKey() < 2) {
                 break;
